Refactor expression functions, fix clampRange.
