+++ src/parsers/expression/encode.js
@@ -1,0 +1,8 @@
+export default function(item, name, retval) {
+  if (item) {
+    var df = this.context.dataflow,
+        target = item.mark.source;
+    df.pulse(target, df.changeset().encode(item, name));
+  }
+  return retval !== undefined ? retval : item;
+}
