+++ src/parsers/expression/prefixes.js
@@ -1,0 +1,3 @@
+export var indexPrefix  = '@';
+export var scalePrefix  = '%';
+export var tuplePrefix  = ':';
