+++ src/parsers/expression/span.js
@@ -1,0 +1,3 @@
+export default function(array) {
+  return array[array.length-1] - array[0];
+}
