+++ src/parsers/expression/tuples.js
@@ -1,0 +1,4 @@
+export default function(name) {
+  var data = this.context.data[name];
+  return data ? data.values.value : [];
+}
